front: correct bug with refresh on resize
