+++ front/src/applications/customget/views/SpaceTimeChart.js
@@ -4,11 +4,7 @@ import React, { useEffect, useRef, useState } from 'react';
 import enableInteractivity, {
   traceVerticalLine,
 } from 'applications/customget/components/enableInteractivity';
-import {
-  handleWindowResize,
-  interpolateOnTime,
-  timeShiftTrain,
-} from 'applications/customget/components/ChartHelpers';
+import { interpolateOnTime, timeShiftTrain } from 'applications/customget/components/ChartHelpers';
 import {
   updateChart,
   updateContextMenu,
@@ -196,11 +192,7 @@ export default function SpaceTimeChart(props) {
     setDataSimulation(createTrain(dispatch, keyValues, simulation.trains, t));
     if (dataSimulation) {
       drawAllTrains(resetChart);
-      handleWindowResize(CHART_ID, dispatch, drawAllTrains, isResizeActive, setResizeActive);
     }
-    return () => {
-      window.removeEventListener('resize');
-    };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [mustRedraw, rotate, selectedTrain, consolidatedSimulation]);
 
@@ -212,11 +204,7 @@ export default function SpaceTimeChart(props) {
       // ADN drawAllTrain already traceVerticalLines
 
       drawAllTrains(resetChart, true, newDataSimulation);
-      handleWindowResize(CHART_ID, dispatch, drawAllTrains, isResizeActive, setResizeActive);
     }
-    return () => {
-      window.removeEventListener('resize');
-    };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [simulation.trains]);
 
@@ -251,10 +239,18 @@ export default function SpaceTimeChart(props) {
   }, [positionValues]);
 
   useEffect(() => {
+    let timeOutFunctionId;
+    const timeOutResize = () => {
+      clearTimeout(timeOutFunctionId);
+      timeOutFunctionId = setTimeout(() => dispatch(updateMustRedraw(true)), 500);
+    };
     window.addEventListener('keydown', handleKey);
+    window.addEventListener('resize', timeOutResize);
     return () => {
       window.removeEventListener('keydown', handleKey);
+      window.removeEventListener('resize', timeOutResize);
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
