+++ front/src/applications/operationalStudies/components/SimulationResults/SpaceTimeChart.js
@@ -5,7 +5,6 @@ import enableInteractivity, {
   traceVerticalLine,
 } from 'applications/operationalStudies/components/SimulationResults/ChartHelpers/enableInteractivity';
 import {
-  handleWindowResize,
   interpolateOnTime,
   timeShiftTrain,
 } from 'applications/operationalStudies/components/SimulationResults/ChartHelpers/ChartHelpers';
@@ -209,11 +208,7 @@ export default function SpaceTimeChart(props) {
     setDataSimulation(createTrain(dispatch, keyValues, simulation.trains, t));
     if (dataSimulation) {
       drawAllTrains(resetChart);
-      handleWindowResize(CHART_ID, dispatch, drawAllTrains, isResizeActive, setResizeActive);
     }
-    return () => {
-      window.removeEventListener('resize');
-    };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [mustRedraw, rotate, selectedTrain, consolidatedSimulation]);
 
@@ -223,13 +218,8 @@ export default function SpaceTimeChart(props) {
     const newDataSimulation = createTrain(dispatch, keyValues, simulation.trains, t);
     if (dataSimulation) {
       // ADN drawAllTrain already traceVerticalLines
-
       drawAllTrains(resetChart, true, newDataSimulation);
-      handleWindowResize(CHART_ID, dispatch, drawAllTrains, isResizeActive, setResizeActive);
     }
-    return () => {
-      window.removeEventListener('resize');
-    };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [simulation.trains]);
 
@@ -262,10 +252,18 @@ export default function SpaceTimeChart(props) {
   }, [positionValues]);
 
   useEffect(() => {
+    let timeOutFunctionId;
+    const timeOutResize = () => {
+      clearTimeout(timeOutFunctionId);
+      timeOutFunctionId = setTimeout(() => dispatch(updateMustRedraw(true)), 500);
+    };
     window.addEventListener('keydown', handleKey);
+    window.addEventListener('resize', timeOutResize);
     return () => {
       window.removeEventListener('keydown', handleKey);
+      window.removeEventListener('resize', timeOutResize);
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
